+++ deltafi-ui/CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 - Added `uploadedBy` field to metadata of all uploaded DeltaFiles (#283)
+- Added support for content segments (#288)
 
 ### Fixed